@@ -5,11 +5,22 @@
   "scripts": {
     "start:cli": "npm --prefix ./packages/cli run start",
     "build:cli": "nx build cli && chmod +x ./dist/packages/cli/bin/index.js",
+    "repo:start": "npm run build:cli && smile-config",
     "postbuild:cli": "cp -R ./packages/cli/configs/ ./dist/packages/cli/configs",
-    "link:cli": "cd ./dist/packages/cli && npm link"
+    "link:cli": "cd ./dist/packages/cli && npm link",
+    "prepare": "husky install",
+    "lint": "npm run lint:branch-name && npm run lint:workspace && npm run affected:lint",
+    "lint:branch-name": "branch-name-lint ./.branchlintrc.json",
+    "lint:commit-name": "commitlint --edit",
+    "lint:js": "eslint ./packages/**/*.ts",
+    "lint:js:fix": "npm run lint:js -- --fix",
+    "lint:workspace": "nx workspace-lint",
+    "affected:lint": "nx affected:lint --parallel --maxParallel=5 --cache",
+    "affected:lint:fix": "nx affected:lint --parallel --maxParallel=5 --cache --fix"
   },
   "private": true,
   "devDependencies": {
+    "@commitlint/config-conventional": "^13.1.0",
     "@nrwl/cli": "12.9.0",
     "@nrwl/eslint-plugin-nx": "12.9.0",
     "@nrwl/jest": "12.9.0",
@@ -18,18 +29,31 @@
     "@nrwl/tao": "12.9.0",
     "@nrwl/workspace": "12.9.0",
     "@types/deep-equal": "^1.0.1",
+    "@types/eslint": "^7.28.0",
     "@types/inquirer": "^8.1.2",
     "@types/jest": "26.0.24",
     "@types/node": "14.14.33",
+    "@types/stylelint": "^13.13.2",
     "@typescript-eslint/eslint-plugin": "~4.28.3",
+    "@typescript-eslint/eslint-plugin-tslint": "^4.31.0",
     "@typescript-eslint/parser": "~4.28.3",
-    "eslint": "7.22.0",
+    "branch-name-lint": "^2.1.1",
+    "commitlint": "^13.1.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "8.1.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.2",
     "jest": "27.0.3",
+    "lint-staged": "^11.1.2",
     "nodemon": "^2.0.13",
     "prettier": "^2.3.1",
+    "smile-track": "^0.17.0",
+    "stylelint": "^13.13.1",
     "ts-jest": "27.0.3",
+    "tslint": "^6.1.3",
     "typescript": "~4.3.5"
   },
   "dependencies": {
@@ -37,6 +61,7 @@
     "inquirer": "^8.1.5",
     "json-merger": "^1.1.6",
     "ts-node": "^9.1.1",
-    "tslib": "^2.0.0"
+    "tslib": "^2.0.0",
+    "type-fest": "^0.21.3"
   }
 }
