@@ -1,38 +1,73 @@
+// eslint-disable-next-line @typescript-eslint/no-require-imports
+import deepEqual = require('deep-equal');
 import type { Linter } from 'eslint';
 import { mergeObjects } from 'json-merger';
-import '@typescript-eslint/experimental-utils';
-import deepEqual = require('deep-equal');
+
+import { NOT_FOUND_INDEX } from '../../const';
 import { coerceArray, mergeArray } from '../utils';
 
 export class EslintMerger {
-  mergeConfigs(target: Linter.Config | null, source: Linter.Config) {
+  mergeConfigs(
+    target: Linter.Config | null,
+    source: Linter.Config
+  ): Linter.Config {
     if (target === null) {
       return source;
     }
 
-    const defaultMerge: Linter.Config = mergeObjects([target, source]);
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+    const defaultMerge: Linter.Config = mergeObjects([
+      target,
+      source,
+    ]) as Linter.Config;
+
+    if (target.plugins && source.plugins) {
+      defaultMerge.plugins = mergeArray(target.plugins, source.plugins);
+    }
 
     if (target.extends && source.extends) {
       defaultMerge.extends = mergeArray(target.extends, source.extends);
     }
 
     if (target.ignorePatterns && source.ignorePatterns) {
-      defaultMerge.ignorePatterns = mergeArray(target.ignorePatterns, source.ignorePatterns);
+      defaultMerge.ignorePatterns = mergeArray(
+        target.ignorePatterns,
+        source.ignorePatterns
+      );
     }
 
     if (target.overrides && source.overrides) {
-      const leftOverrides = JSON.parse(JSON.stringify(source.overrides)) as Linter.ConfigOverride[];
+      const leftOverrides = JSON.parse(
+        JSON.stringify(source.overrides)
+      ) as Linter.ConfigOverride[];
 
       defaultMerge.overrides = [
         ...target.overrides.map((targetOverride) => {
-          const foundSourceOverrideIndex = leftOverrides.findIndex((sourceOverride) => deepEqual(coerceArray(targetOverride.files).slice().sort(), coerceArray(sourceOverride.files).slice().sort()))
+          const foundSourceOverrideIndex = leftOverrides.findIndex(
+            (sourceOverride) =>
+              deepEqual(
+                coerceArray(targetOverride.files)
+                  .slice()
+                  .sort((a, b) => a.localeCompare(b)),
+                coerceArray(sourceOverride.files)
+                  .slice()
+                  .sort((a, b) => a.localeCompare(b))
+              )
+          );
 
-          if (foundSourceOverrideIndex !== -1) {
+          if (foundSourceOverrideIndex !== NOT_FOUND_INDEX) {
             const foundSourceOverride = leftOverrides[foundSourceOverrideIndex];
-            const mergedOverride = mergeObjects([targetOverride, foundSourceOverride]) as Linter.ConfigOverride;
-            mergedOverride.plugins = mergeArray(targetOverride.plugins, foundSourceOverride.plugins);
+            const mergedOverride = mergeObjects([
+              targetOverride,
+              foundSourceOverride,
+            ]) as Linter.ConfigOverride;
+            mergedOverride.plugins = mergeArray(
+              targetOverride.plugins,
+              foundSourceOverride.plugins
+            );
 
-            leftOverrides.splice(foundSourceOverrideIndex, 1);
+            const DELETE_COUNT = 1;
+            leftOverrides.splice(foundSourceOverrideIndex, DELETE_COUNT);
 
             return mergedOverride;
           }
