@@ -1,20 +1,23 @@
 #!/usr/bin/env node
 import { prompt } from 'inquirer';
+
 import { builtInConfigs } from '../src';
+import { FIRST_INDEX, MORE_THEN_ONE } from '../src/const';
 
 (async () => {
+  let config = builtInConfigs[FIRST_INDEX];
 
-  let config = builtInConfigs[0];
-
-  if (builtInConfigs.length > 1) {
+  if (builtInConfigs.length > MORE_THEN_ONE) {
     config = await prompt({
       type: 'list',
       name: 'configIndex',
       message: 'Which style guide do you want to follow?',
-      choices: builtInConfigs.map((config, index) => ({
-        name: `${config.name}: ${config.url}`,
+      choices: builtInConfigs.map((choiceConfig, index) => ({
+        name: `${choiceConfig.name}: ${choiceConfig.url}`,
         value: String(index),
       })),
     }).then((value: Record<any, any>) => builtInConfigs[+value.configIndex]);
   }
-})();
+})().catch((error) => {
+  throw error;
+});
