@@ -1,13 +1,17 @@
-import { BaseConfigItemModule } from '../../../../src/base';
-import { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
+import type { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
 
-export class BranchNameLintModule extends BaseConfigItemModule implements AbstractConfigItemModule {
+import { BaseConfigItemModule } from '../../../../src/base';
+
+export class BranchNameLintModule
+  extends BaseConfigItemModule
+  implements AbstractConfigItemModule
+{
   name = 'branch-name-lint';
   includeToLintScript = [
     {
       command: 'lint:branch-name',
       order: 0,
-    }
+    },
   ];
 
   constructor() {
