@@ -1,7 +1,11 @@
-import { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
+import type { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
+
 import { BaseConfigItemModule } from '../../../../src/base';
 
-export class LintStagedModule extends BaseConfigItemModule implements AbstractConfigItemModule {
+export class LintStagedModule
+  extends BaseConfigItemModule
+  implements AbstractConfigItemModule
+{
   name = 'lint-staged';
 
   constructor() {
