@@ -1,9 +1,12 @@
-import { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
+import type { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
 
 import { BaseConfigItemModule } from '../../../../src/base';
 import { prettierAddons } from './addons';
 
-export class PrettierModule extends BaseConfigItemModule implements AbstractConfigItemModule {
+export class PrettierModule
+  extends BaseConfigItemModule
+  implements AbstractConfigItemModule
+{
   name = 'prettier';
   addons = prettierAddons;
 
