@@ -1,7 +1,11 @@
-import { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
+import type { AbstractConfigItemModule } from '@smile-config/cli/interfaces';
+
 import { BaseConfigItemModule } from '../../../../src/base';
 
-export class SmileTrackModule extends BaseConfigItemModule implements AbstractConfigItemModule {
+export class SmileTrackModule
+  extends BaseConfigItemModule
+  implements AbstractConfigItemModule
+{
   name = 'smile-track';
 
   constructor() {
