@@ -1,6 +1,6 @@
 export class NoRequiredFileError extends Error {
   name = 'NoRequiredFileError';
-  message = 'There is no config required file in the folder: "' + this.requiredFile + '"';
+  message = `There is no config required file in the folder: "${this.requiredFile}"`;
 
   constructor(public requiredFile: string) {
     super();
