@@ -1,4 +1,7 @@
+/* eslint-disable class-methods-use-this */
 import * as fs from 'fs';
+
+import { FIRST_INDEX, ONE_ITEM_LENGTH, SPLICE_LAST_ELEMENT } from '../../const';
 
 export enum FileType {
   JSON = 'json',
@@ -13,16 +16,20 @@ export class FolderService {
   }
 
   getFileName(path: string) {
-    return path.split('files/').splice(-1)[0];
+    return path.split('files/').splice(SPLICE_LAST_ELEMENT)[FIRST_INDEX];
   }
 
   isNestedFile(path: string): boolean {
-    return path.split('/').length !== 1;
+    return path.split('/').length !== ONE_ITEM_LENGTH;
   }
 
   getFileType(path: string): FileType | string {
-    const file = path.split('/').splice(-1)[0];
-    const type = path.split('/').splice(-1)[0].split('.').splice(-1)[0];
+    const file = path.split('/').splice(SPLICE_LAST_ELEMENT)[FIRST_INDEX];
+    const type = path
+      .split('/')
+      .splice(SPLICE_LAST_ELEMENT)
+      [FIRST_INDEX].split('.')
+      .splice(SPLICE_LAST_ELEMENT)[FIRST_INDEX];
 
     return type === file ? FileType.NO_EXTENSION : type;
   }
